requirements.txt
amqp
celery
Django
django-cors-headers
django-filter
djangorestframework
drf-spectacular
joblib
numpy
packaging
pandas
pillow
prompt_toolkit
pyparsing
python-dateutil
python-magic
pytz
PyYAML
redis
referencing
reportlab
rpds-py
scapy
scikit-learn
scipy
six
sqlparse
